login and regist for costumers working

+++ ClientApp/app/src/main/java/pt/up/fe/cmov16/client/clientapp/database/ProductContract.java
@@ -111,7 +111,7 @@ public final class ProductContract {
                             ProductEntry.COLUMN_NAME_UNIT_PRICE));
 
                     Product prod = new Product();
-                    prod.setUnitprice(Double.valueOf(unitPrice));
+                    prod.setUnitprice(Float.valueOf(unitPrice));
                     prod.setName(name);
                     products.add(prod);
                     // move to next row
